+++ app/src/main/java/com/sagrd/travelmanagement/data/ViajeDao.kt
@@ -15,9 +15,9 @@ interface ViajeDao {
     @Update
     suspend fun Update(viaje: Viaje)
 
-    @Query("SELECT * FROM Viaje WHERE ViajeId= :key")
+    @Query("SELECT * FROM Viaje WHERE travelId= :key")
      fun Find(key:Long): LiveData<Viaje>
 
-    @Query("SELECT * FROM Viaje ORDER BY ViajeId DESC")
+    @Query("SELECT * FROM Viaje ORDER BY travelId DESC")
      fun Lista() :LiveData<List<Viaje>>
 }
