+++ app/src/main/java/com/sagrd/travelmanagement/network/RetrofitInstance.kt
@@ -11,7 +11,6 @@ object RetrofitInstance {
             .addConverterFactory(GsonConverterFactory.create())
             .build()
             .create(TravelApi::class.java)
-
     }
 
 }
